+++ test/HylianShield/Tests/Validator/Number/NegativeTest.php
@@ -4,12 +4,9 @@
  *
  * @package HylianShield
  * @subpackage Test
- * @copyright 2014 Jan-Marten "Joh Man X" de Boer
  */
 
 namespace HylianShield\Tests\Validator\Number;
-
-use \HylianShield\Validator\Number\Negative;
 
 /**
  * Negative test.