+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # v0.8.0
 
 - Added an interface for validators
+- Deprecated method type on Validator in favor of getType
 - Added context to validations, which are used to enrich the messages of 
 validators.
-- Started support for HHVM
-- Added support for PHP 5.6
+- Started support for HHVM through Travis CI
+- Added support for PHP 5.6 through Travis CI
+- Fixed code style issues
+- Removed number tests using octet literals, as PHP does not interpret them 
+as expected.
 
 # v0.7.0
 
@@ -39,7 +43,7 @@ validators.
 
 # v0.2.2
 
-- Fix the numeric arrar validator unit test
+- Fix the numeric array validator unit test
 - Improve performance of Range validators
 
 # v0.2.1